% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_renv_datapkg.R
\name{bb_renv_datapkg}
\alias{bb_renv_datapkg}
\title{Install Or Update A Local R Data Package}
\usage{
bb_renv_datapkg(path)
}
\arguments{
\item{path}{Path to a directory containing one or more versions of the same data package.}
}
\value{
noting
}
\description{
If a directory is specified, this function will compare the currently installed version to the latest available version in the directory.  If there is a newer version available (based on version number), it will install this version.  If a tarball is specifically requested it will install that one. If the package hasn't ever been installed, it will install the newest version.
}
\seealso{
\code{\link[purrr]{safely}},\code{\link[purrr]{map}}
\code{\link[stringr]{str_detect}},\code{\link[stringr]{str_glue}},\code{\link[stringr]{str_split}},\code{\link[stringr]{str_replace}},\code{\link[stringr]{str_sub}}
\code{\link[renv]{install}}
\code{\link[tibble]{as_tibble}}
\code{\link[dplyr]{filter}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{slice}},\code{\link[dplyr]{pull}}
}
