% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_2.0.R
\name{cache_fun}
\alias{cache_fun}
\title{cache a single package}
\usage{
cache_fun(
  package,
  cache_loc = fs::path(Sys.getenv("BLASERTEMPLATES_CACHE_ROOT"), "library")
)
}
\description{
cache a single package
}
