% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_2.0.R
\name{catch_blasertemplates_root}
\alias{catch_blasertemplates_root}
\title{Make sure you are in a properly-formatted blaseRtemplates project.}
\usage{
catch_blasertemplates_root()
}
\description{
Make sure you are in a properly-formatted blaseRtemplates project.
}
