% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_functions.R
\name{easy_init}
\alias{easy_init}
\title{Easily Initiate an renv Project}
\usage{
easy_init()
}
\value{
Nothing
}
\description{
Behave similarly to renv::init but instead of downloading and building everything, will first attempt to link to the latest versions of packages available in your cache.
}
\seealso{
\code{\link[renv]{init}}, \code{\link[renv]{dependencies}}
\code{\link[purrr]{map}}, \code{\link[purrr]{safely}}
}
