% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_functions.R
\name{easy_restore}
\alias{easy_restore}
\title{Easily restore a project from an RENV Lockfile}
\usage{
easy_restore(lockfile = "default")
}
\arguments{
\item{lockfile}{Optional non-default lock file to restore from.  Otherwise will use "renv.lock", Default: 'default'}
}
\value{
Nothing
}
\description{
Reads the project lockfile or another lockfile supplied as an argument.  Then attempts to link to a binary version of each package in the cache.  If unavailable it will attempt to link to the newest version in the cache.  If that is unavailable it will install the package from the available repositories.
}
\seealso{
\code{\link[rjson]{fromJSON}}
\code{\link[purrr]{map}}
}
