% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.R
\name{fix_another_library}
\alias{fix_another_library}
\title{Fix a User-Project Library}
\usage{
fix_another_library(file, dir)
}
\arguments{
\item{file}{The library catalog tsv file to read from.}

\item{dir}{The user project library to repair.}
}
\value{
Nothing
}
\description{
Sometimes, the user_project library can break.  This can happen if there are failures in the upstream install functions.  Or if the links are deleted for some reason.  If everything is broken, you may need to repair the user project library.  To do this, exit the project and enter a working project.  Delete the links in the offending user project library.  Then run this function to relink.  Use either the library catalog file from the project itself, or if this is also corrupted with bad information, run an older version that worked or a version from another project that worked.  If you can identify the problematic package in the course of these fixes, then you should probably delete it from your cache entirely.
}
\seealso{
\code{\link[fs]{path_math}}, \code{\link[fs]{create}}, \code{\link[fs]{path}}, \code{\link[fs]{path_file}}
\code{\link[readr]{read_delim}}
\code{\link[dplyr]{pull}}
}
