% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{fork_github_project}
\alias{fork_github_project}
\title{Fork A Project From Github}
\usage{
fork_github_project(repo, dest = NULL)
}
\arguments{
\item{repo}{Repository name in the form of <owner>/<repo> or a url}

\item{dest}{Parent directory in which to create the forked project, Default: NULL}
}
\description{
This wraps usethis::create_from_github and adds 2 templates which are not tracked by git and are missed by the base function:  git_commands.R, local_configs.R.  This also enforces fork = TRUE and opens the project in a new session. You will need to have write access to the originator's repo for this to work correctly and you should provide write access for the originator to your repo for them to be able to update your shared work directly.
}
\seealso{
\code{\link[usethis]{create_from_github}},\code{\link[usethis]{use_template}},\code{\link[usethis]{proj_activate}}
\code{\link[stringr]{str_replace}},\code{\link[stringr]{str_extract}}
}
