% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_2.0.R
\name{get_new_library}
\alias{get_new_library}
\title{Get A New Project Library}
\usage{
get_new_library(newest_or_file = "newest")
}
\arguments{
\item{newest_or_file}{Which set of packages to symlink, Default: 'newest'}
}
\value{
Uninstalled packages hashes.
}
\description{
Use this to replace the current symlinked library with a new version.  By default, the function will link to the newest version of all packages available in the cache.  Alternatively, identify another project library catalog to replace the current version.
}
