% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_functions.R
\name{git_easy_branch}
\alias{git_easy_branch}
\title{Easily Create or Switch Git Branches}
\usage{
git_easy_branch(branch)
}
\arguments{
\item{branch}{A character string with the branch name to create or switch to.}
}
\value{
nothing
\code{\link[gert]{git_branch}}
}
\description{
Supply this function with a branch name.  If the branch exists it will switch to the branch.  If not, it will pull any changes from remote and then create the branch.  Any uncommitted work will be carried over to the new branch in the same state.  Avoid repeatedly switching branches with work in different states of completion since this may cause conflicts
}
