% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_functions.R
\name{git_push_all}
\alias{git_push_all}
\title{Push to All Remotes}
\usage{
git_push_all()
}
\value{
nothing
}
\description{
This uses gert to look up all active remotes and then runs gert::git_push() to each.
}
