% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_functions.R
\name{git_rewind_to}
\alias{git_rewind_to}
\title{Rewind Git History}
\usage{
git_rewind_to(commit)
}
\arguments{
\item{commit}{Hash of the commit you want to rewind the state of your repository to.  Requires a minimum of 7 characters.}
}
\value{
a tibble with the new git commit log after rewinding
}
\description{
This function uses git revert to rewind history to a prior commit.  First make sure all of your changes have been committed.  Then run gert::git_log() to identify the "good" commit you want to rewind to.  Supply this as the argument to this function.  A new commit will be made with a helpful message.  Commit history is not changed so you can always rewind the rewind etc....
}
\seealso{
\code{\link[gert]{git_commit}}
}
