% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_functions.R
\name{git_safe_merge}
\alias{git_safe_merge}
\title{Safely Merge your Working Branch}
\usage{
git_safe_merge(branch = NULL, upstream = "main")
}
\arguments{
\item{branch}{The working branch you wish to merge, Default: NULL}

\item{upstream}{The default upstream branch you wish to merge into, Default: NULL}
}
\value{
nothing
}
\description{
This function updates default branch (usually "main") from remote.  This pulls in any changes from other contributors.  Then it merges the working branch into the upstream branch.
}
\seealso{
\code{\link[usethis]{git-default-branch}}
\code{\link[gert]{git_branch}},\code{\link[gert]{git_commit}}
}
