% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_functions.R
\name{git_update_continue}
\alias{git_update_continue}
\title{Continue Updating Branch after Resolving Conflict}
\usage{
git_update_continue()
}
\value{
nothing
}
\description{
This function adds files after conflict resolution and continues the rebase.
}
\seealso{
\code{\link[gert]{git_commit}}
}
