% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_functions.R
\name{gitcreds_set}
\alias{gitcreds_set}
\title{Interactively Set Github PAT}
\usage{
gitcreds_set()
}
\value{
nothing
}
\description{
This wraps gitcreds::gitcreds_set() and adds a system call to edit the user global git config to set the cache timeout to 1 billion seconds if running a Linux system.
}
\seealso{
\code{\link[gitcreds]{gitcreds_get}}
}
