% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_2.0.R
\name{hash_n_cache}
\alias{hash_n_cache}
\title{hash one or more functions and then cache them and update the catalogs}
\usage{
hash_n_cache()
}
\description{
hash one or more functions and then cache them and update the catalogs
}
