% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_2.0.R
\name{initialize_github}
\alias{initialize_github}
\title{Initialize A Project By Forking A Github Repo}
\usage{
initialize_github(repo, dest, library = "newest", open = FALSE)
}
\arguments{
\item{repo}{The repo to clone.  Must be in the form of github_user/repo_name.  If private, you must be a collaborator and have permission to fork the repo from the owner.}

\item{dest}{Destination directory.  This directory will become the parent directory for the project you are forking.}

\item{library}{Optional blaseRtemplates library catalog to use when making the project package library.  If nothing is entered, the newest available versions will be installed.  Otherwise, enter a library catalog file in the form of "library_catalogs/user_repo_name.tsv.  Of course replace user and repo_name.  This file will have to exist in the project you are forking., Default: 'newest'}

\item{open}{Whether to open the forked project, Default: FALSE}
}
\value{
nothing
}
\description{
This function wraps usethis::create_from_github, making some useful default choices and building the user project library compatible with the blaseRtemplates R system.  Because this function forks the project, git will set up the originator as an upstream remote.  Using blaseRtemplates::git_push_all will push to both the originator and the collaborator's github.
}
\seealso{
\code{\link[stringr]{str_remove}}
\code{\link[cli]{cli_abort}}
\code{\link[usethis]{create_from_github}}, \code{\link[usethis]{proj_utils}}, \code{\link[usethis]{proj_activate}}
\code{\link[fs]{create}}, \code{\link[fs]{path}}
\code{\link[withr]{defer}}
}
