% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_2.0.R
\name{initialize_package}
\alias{initialize_package}
\title{Initialize a Package Using a Standard Template}
\usage{
initialize_package(path, fields = list(), roxygen = TRUE, check_name = TRUE)
}
\arguments{
\item{path}{path/name for the new package.  It should include letters and "." only to be CRAN-compliant.}

\item{fields}{named list of fields in addition to/overriding defaults for the DESCRIPTION file, Default: list()}

\item{roxygen}{do you plan to use roxygen2 to document package?, Default: TRUE}

\item{check_name}{check if name is CRAN-compliant, Default: TRUE}

\item{rstudio}{makes an Rstudio project, default is true}

\item{open}{to open or not, Default: rlang::is_interactive()}
}
\description{
This wraps \code{usethis::create_package()} and adds a few additional templates.
}
\seealso{
\code{\link[rstudioapi]{isAvailable}}
\code{\link[rlang]{is_interactive}}
\code{\link[usethis]{use_template}}
\code{\link[withr]{defer}}
}
