% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_2.0.R
\name{rec_get_deps_internal}
\alias{rec_get_deps_internal}
\title{recursively get package dependencies}
\usage{
rec_get_deps_internal(needed, checked, deps, catalog)
}
\description{
recursively get package dependencies
}
