% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish.R
\name{regenerate_bt_configs}
\alias{regenerate_bt_configs}
\title{Regenerate .Rprofile and .Renviron Files}
\usage{
regenerate_bt_configs(cache_path, project_path)
}
\arguments{
\item{cache_path}{path to the cache directory}

\item{project_path}{path to the projects directory}
}
\value{
nothing
}
\description{
If you have deleted or otherwise broken your .Rprofile or .Renviron files, you may have difficulty connecting to the package cache.  This function will regenerate both for you.  The existing .Rprofile must be deleted manually.  You can choose to archive the old version if you wish.  It will be replaced with the standard .Rprofile from blaseRtemplates.  The .Renviron file will be modified by removing the damaged lines and replacing them with the correct ones.  You must supply the correct file path locations to your cache directory and project directory, otherwise your R installation will be configured incorrectly.
}
\seealso{
\code{\link[fs]{file_access}}, \code{\link[fs]{path}}, \code{\link[fs]{copy}}, \code{\link[fs]{path_package}}
\code{\link[cli]{cli_abort}}, \code{\link[cli]{cli_alert}}
\code{\link[stringr]{str_detect}}, \code{\link[stringr]{str_replace}}
}
