% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_functions.R
\name{regenerate_git_commands}
\alias{regenerate_git_commands}
\title{Regenerate a Git Commands File}
\usage{
regenerate_git_commands()
}
\value{
nothing
}
\description{
Since this file is ignored by git, you will have to regenerate it when forking a repository.  This function writes the template file to your R directory as "regenerated_git_commands.R".
}
