% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_functions.R
\name{reset_prompt}
\alias{reset_prompt}
\title{Reset Your Console Prompt}
\usage{
reset_prompt()
}
\value{
nothing
}
\description{
The prompt package adds a nice feature but has some limitations, namely, that it does not respect changing git branches and has to be manually re-called.  This defeats the purpose.  Blasertemplates git functions automoatically call prompt to change the prompt label when switching branches, but this will not happen if you change branches using the terminal, the git panel or other git branching functions.  Therefore this function is provided to manually reset your prompt to the current branch.
}
\seealso{
\code{\link[prompt]{set_prompt}},\code{\link[prompt]{prompt_git}}
}
