% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_git_collab.R
\name{setup_git_collab}
\alias{setup_git_collab}
\title{Install Scripts to Support Git Collaboration}
\usage{
setup_git_collab()
}
\value{
nothing
}
\description{
This function installs two scripts that bundle git commands to help avoid conflicts.  gitupdatebranch switches from working branch to main, updates from remote, then rebases working branch.  Git safemerge swtiches to main and pulls before merging.  Both scripts are written to ~/.local/bin and the global git config is edited to provide these aliases:  git updatebranch, git safemerge.
}
