% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_functions.R
\name{sync_cache}
\alias{sync_cache}
\title{Synchronize Project Library with Renv Cache}
\usage{
sync_cache()
}
\value{
nothing
}
\description{
This function snapshots your current library, copies any packages existing in the renv project library to the renv cache and then links the package back into the project library.  This leaves you with an updated cache and a project library containing only symlinks, which is usually the ideal status for your project.
}
\seealso{
\code{\link[renv]{snapshot}},\code{\link[renv]{paths}},\code{\link[renv]{character(0)}}
}
