% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_2.0.R
\name{update_dependency_catalog}
\alias{update_dependency_catalog}
\title{update the dependency catalog}
\usage{
update_dependency_catalog(
  cache_loc = Sys.getenv("BLASERTEMPLATES_CACHE_ROOT"),
  dep_update
)
}
\description{
update the dependency catalog
}
