% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_2.0.R
\name{update_package_catalog}
\alias{update_package_catalog}
\title{update the package catalog}
\usage{
update_package_catalog(
  cache_loc = Sys.getenv("BLASERTEMPLATES_CACHE_ROOT"),
  pkg_update
)
}
\description{
update the package catalog
}
