% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{upgrade_bt}
\alias{upgrade_bt}
\title{Upgrade a Project to Use the Latest Version of BlaseRtemplates}
\usage{
upgrade_bt(
  path = fs::path_wd(),
  path_to_cache_root = Sys.getenv("BLASERTEMPLATES_CACHE_ROOT")
)
}
\arguments{
\item{path}{path to the project to upgrade.  By default will upgrade the current project, Default: fs::path_wd()}

\item{path_to_cache_root}{PARAM_DESCRIPTION, Default: Sys.getenv("BLASERTEMPLATES_CACHE_ROOT")}
}
\description{
This function converts an renv-based project to a blaseRtemplates project using the new cache structure.  Specifically, it will take the following actions on the provided project:
\enumerate{
\item delete the renv folder
\item delete the renv lock file
\item write a new .Rprofile which in turn will
\item create a new user_project directory.  This will be the primary package library for the project.  It will contain symlinks pointing to the newest versions of packages in the main cache.
\item write a new file listing the packages actively used by the project into a new "library_catalogs" directory
}
}
