% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_2.0.R
\name{write_project_library_catalog}
\alias{write_project_library_catalog}
\title{Write A New Project Library Catalog}
\usage{
write_project_library_catalog()
}
\value{
returns nothing
}
\description{
In the current version of blaseRtemplates, the package library is cached at the location designated by the environment variable "BLASERTEMPLATES_CACHE_ROOT".  There is a single cache for all users and projects.  The cache holds the binary software used by each package.  The packages for each project are connected to the cache by symlinks. The cache is versioned so that different projects can use different versions if desired.  Use this function to write a tab-delimited file listing the packages used by each project.

This file will be written to the "library_catalogs" directory within each project.  The filename incorporates the user name so everyone working on the project will have their own.  Use \code{get_new_library()} to adopt a new version of all packages
}
