% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dratify.R
\name{dratify}
\alias{dratify}
\title{Build a Package and Send to Local CRAN-like Repository}
\usage{
dratify(repo_name, repo_dir)
}
\arguments{
\item{repo_name}{Name for the repository.  Check to see if it has been added using options("repos")}

\item{repo_dir}{Directory to send the package source file to.  On linux at least, this can start with "~".}
}
\value{
nothing
}
\description{
Use this to build a package that can be installed like a CRAN package.  This wraps around function from the drat package.  A source package is built, inserted into the local drat repository, and the user session is checked to be sure the designated repository is available.  If not, it modifies the main .Rprofile file to make it accessible.  In order for this to work, the file path must be accessible on the local network.
}
\seealso{
\code{\link[devtools]{build}}
\code{\link[drat]{insertPackage}}
}
