% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_functions.R
\name{git_update_branch}
\alias{git_update_branch}
\title{Update a Working Git Branch}
\usage{
git_update_branch(branch = NULL, upstream = NULL)
}
\arguments{
\item{branch}{The working branch you wish to update, Default: NULL}

\item{upstream}{The default upstream branch you wish to update from, Default: NULL}
}
\value{
nothing
}
\description{
This function updates a git branch via rebase from a default upstream branch (usually "main").  You can explicitly provide the names of your working branch and the default upstream branch.  If not provided, the function will use the current branch as your working branch and will automatically identify the default upstream branch.  Internally this calls a system file that must be installed using blaseRtemplates::setup_git_collab().  The upstream branch also needs to be connected to a remote (e.g. github).
}
\seealso{
\code{\link[usethis]{git-default-branch}}
\code{\link[gert]{git_branch}},\code{\link[gert]{git_stash}}
}
