% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_2.0.R
\name{initialize_github}
\alias{initialize_github}
\title{Initialize A Project By Forking A Github Repo}
\usage{
initialize_github(repo, dest = NULL, open = TRUE)
}
\arguments{
\item{repo}{The repo to clone.  Must be in the form of github_user/repo_name.  If private, you must be a collaborator and have permission to fork the repo from the owner.}

\item{dest}{Destination directory.  This directory will become the parent directory for the project you are forking. If NULL, the default, it will put the project in the directory defined by the usethis.destdir option.  Set this in ~/.Rprofile.}

\item{open}{Whether to open the forked project, Default: TRUE}
}
\value{
nothing
}
\description{
This function wraps usethis::create_from_github, making some useful default choices. Because this function forks the project, git will set up the originator as an upstream remote.  Using blaseRtemplates::git_push_all will push to both the originator and the collaborator's github.
}
