% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R, R/initialize_2.0.R
\name{initialize_package}
\alias{initialize_package}
\title{Initialize a Package Using a Standard}
\usage{
initialize_package(
  path,
  path_to_cache_root = "/workspace/rst/cache_R_4_2",
  fields = list(),
  rstudio = rstudioapi::isAvailable(),
  roxygen = TRUE,
  check_name = TRUE,
  open = rlang::is_interactive()
)

initialize_package(
  path,
  path_to_cache_root = "/workspace/rst/cache_R_4_2",
  fields = list(),
  rstudio = rstudioapi::isAvailable(),
  roxygen = TRUE,
  check_name = TRUE,
  open = rlang::is_interactive()
)
}
\arguments{
\item{path}{path/name for the new package.  It should include letters and "." only to be CRAN-compliant.}

\item{fields}{named list of fields in addition to/overriding defaults for the DESCRIPTION file, Default: list()}

\item{rstudio}{makes an Rstudio project, default is true}

\item{roxygen}{do you plan to use roxygen2 to document package?, Default: TRUE}

\item{check_name}{check if name is CRAN-compliant, Default: TRUE}

\item{open}{to open or not, Default: rlang::is_interactive()}
}
\description{
This wraps usethis::create_package() and adds a few additional templates.

This wraps usethis::create_package() and adds a few additional templates.
}
\seealso{
\code{\link[rstudioapi]{isAvailable}}
\code{\link[rlang]{is_interactive}}
\code{\link[usethis]{use_template}}
\code{\link[withr]{defer}}

\code{\link[rstudioapi]{isAvailable}}
\code{\link[rlang]{is_interactive}}
\code{\link[usethis]{use_template}}
\code{\link[withr]{defer}}
}
