% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_2.0.R
\name{install_one_package}
\alias{install_one_package}
\title{Install One Package}
\usage{
install_one_package(
  package,
  which_version = NULL,
  which_hash = NULL,
  how = c("ask", "new_or_update", "link_from_cache", "tarball")
)
}
\arguments{
\item{package}{PARAM_DESCRIPTION}

\item{which_version}{PARAM_DESCRIPTION, Default: NULL}

\item{which_hash}{PARAM_DESCRIPTION, Default: NULL}

\item{how}{PARAM_DESCRIPTION, Default: c("ask", "new_or_update", "link_from_cache", "tarball")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[cli]{cli_div}}, \code{\link[cli]{cli_alert}}
\code{\link[stringr]{str_detect}}, \code{\link[stringr]{str_replace}}
\code{\link[pak]{pkg_install}}
}
