% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_2.0.R
\name{install_one_package}
\alias{install_one_package}
\title{Install One Package}
\usage{
install_one_package(
  package,
  how = c("ask", "new_or_update", "link_from_cache", "tarball"),
  which_version = NULL,
  which_hash = NULL
)
}
\arguments{
\item{package}{Package name or path to tarball.  Prefix with "repo\/" for github source packages and "bioc::" for bioconductor.}

\item{how}{How to install the package, Default: c("ask", "new_or_update", "link_from_cache", "tarball")}

\item{which_version}{Package version to install, Default: NULL}

\item{which_hash}{Package hash to install, Default: NULL}
}
\value{
nothing
}
\description{
Use this to install a new package.  Choosing "new_or_update" will go to the package repository and get the latest version of the software.  Choosing "link_from_cache" will get you the latest version in the cache, for example if another user has added a new package you want, but you don't want to update the whole library.  Also, use this option with either "which_version" or "which_hash" to install specific versions.
}
