% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_2.0.R
\name{project_data}
\alias{project_data}
\title{Activate Project Data}
\usage{
project_data(path)
}
\arguments{
\item{path}{Path to a directory containing a data package or to a specific data pacakge.}
}
\value{
none
}
\description{
Use this to update, install and/or load project data.  Usual practice is to provide the path to a directory holding data package tarballs.  This function will find the newest version, compare that to the versions in the cache and used in the package and give you the newest version.  Alternatively, provide the path to a specific .tar.gz file to install and activate that one.

After installing, the package will be loaded into a hidden environment using \code{lazyData::requireData()}.  This loads the project into memory only when called.
}
