% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_2.0.R
\name{write_project_library_catalog}
\alias{write_project_library_catalog}
\title{FUNCTION_TITLE}
\usage{
write_project_library_catalog()
}
\arguments{
\item{lib_loc}{PARAM_DESCRIPTION, Default: cache_paths$user_project}

\item{cache_loc}{PARAM_DESCRIPTION, Default: cache_paths$library}

\item{user}{PARAM_DESCRIPTION, Default: Sys.getenv()[\link{"USER"}]}

\item{project}{PARAM_DESCRIPTION, Default: fs::path_file(usethis::proj_get())}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[fs]{path_file}}, \code{\link[fs]{dir_ls}}, \code{\link[fs]{is_file}}, \code{\link[fs]{link_path}}, \code{\link[fs]{create}}, \code{\link[fs]{path}}
\code{\link[usethis]{proj_utils}}
\code{\link[purrr]{map}}
\code{\link[tibble]{tibble}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{case_when}}
\code{\link[renv]{dependencies}}
\code{\link[readr]{read_delim}}, \code{\link[readr]{write_delim}}
}
